add root folder

+++ PM.WebAPI/Services/FolderService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using PM.Common.Models.Rest;
 using PM.Data.Entities;
@@ -14,36 +15,42 @@ namespace PM.WebAPI.Services
     {
         private readonly IFolderRepository _folderRepository;
         private readonly IMapper _mapper;
+        private readonly UserManager<ApplicationUser> _userManager;
 
         public FolderService(IFolderRepository folderRepository,
-            IMapper mapper)
+            IMapper mapper,
+            UserManager<ApplicationUser> userManager)
         {
             _folderRepository = folderRepository;
             _mapper = mapper;
+            _userManager = userManager;
         }
 
-        public FolderRestModel GetTree(string userId)
+        public async Task<FolderRestModel> GetTreeAsync(string userId)
         {
             var folders = _folderRepository.GetQueryable()
                 .Include(x => x.ChildFolders)
                 .Where(x => x.CreatorId == userId)
                 .ToList();
 
-            var rootFolder = CreateTree(folders, userId);
+            var rootFolder = await GetTreeRoot(userId, folders);
+
             return _mapper.Map<FolderRestModel>(rootFolder);
         }
 
         public async Task<FolderRestModel> CreateFolderAsync(string userId, FolderRestModel folderRest)
         {
-            if (folderRest.ParentFolderId != null)
+            if (folderRest.ParentFolderId == null)
             {
-                var fatherFolder = _folderRepository
-                        .GetList(x => x.CreatorId == userId && x.Id == folderRest.ParentFolderId);
+                return null;
+            }
 
-                if (fatherFolder.Count != 1)
-                {
-                    return null;
-                }
+            var fatherFolder = _folderRepository
+                    .GetList(x => x.CreatorId == userId && x.Id == folderRest.ParentFolderId);
+
+            if (fatherFolder.Count != 1)
+            {
+                return null;
             }
 
             var folder = _mapper.Map<Folder>(folderRest);
@@ -57,7 +64,7 @@ namespace PM.WebAPI.Services
             var folder = _folderRepository
                 .GetList(x => x.CreatorId == userId && x.Id == id).FirstOrDefault();
 
-            if (folder == null)
+            if (folder == null || folder.ParentFolderId == null)
             {
                 return null;
             }
@@ -69,14 +76,29 @@ namespace PM.WebAPI.Services
 
         public async Task DeleteAsync(int id, string userId)
         {
+            var folder = _folderRepository
+                .GetList(x => x.CreatorId == userId && x.Id == id).FirstOrDefault();
+
+            if (folder == null || folder.ParentFolderId == null)
+            {
+                return;
+            }
+
             await _folderRepository.DeleteFolderAsync(id, userId);
         }
 
-        private Folder CreateTree(IList<Folder> folders, string userId)
+        private async Task<Folder> GetTreeRoot(string userId, List<Folder> folders)
         {
-            var rootFolder = new Folder { Name = "", CreatorId = userId };
-            rootFolder.ChildFolders = folders.Where(x => x.ParentFolderId == null).ToList();
-            return rootFolder;
+            if (folders.Count() == 0)
+            {
+                var currentUser = await _userManager.FindByIdAsync(userId);
+                var rootFolder = new Folder { Name = "root", CreatorId = userId };
+                currentUser.RootFolder = rootFolder;
+                rootFolder = await _folderRepository.SaveAsync(rootFolder);
+                folders.Add(rootFolder);
+            }
+
+            return folders.FirstOrDefault(x => x.ParentFolderId == null);
         }
     }
 }
